Update documentation
This commit updates the build documentation
and provides base-line minimalist example kernel
configs that are highly optimized for RTAI, and to
give both new users and experienced developers a rough
idea on how RTAI should be configured in the kernel.

A few more commits may need to take place on top of
this one, in order to sort out any mistakes in this commit.

Signed-off-by: Alec Ari <neotheuser@ymail.com>
Signed-off-by: Shahbaz Youssefi <ShabbyX@gmail.com>