rtai-config stub script:  add `-o` arg to select the 'only' headers found
(Fixup for ShabbyX/RTAI#16)

If only one set of headers is found on the system, the `-o` option
selects this kernel to build against.  If more than one is found,
error out.

The use case for this is on a build system where the running kernel is
not an RTAI kernel, where exactly one set of RTAI kernel headers
should be installed, and where it would be an irritating exercise to
ask the build scripts to repeat many of the checks found in this
script.  ;)