Fix `make distclean`:  Create `Makefile` symlinks for all arches
The `distclean` target runs the kbuild `clean` target in *each
subdirectory* in `$(DIST_SUBDIRS)`, and that requires `Makefile` to
exist.  The `Makefile` in each subdir is a symlink created by
`./configure`, which only created a symlink for the target arch and
not other arches.  This caused `distclean` to fail when running the
kbuild `make clean` in non-target arch directories.

This commit simply adds symlinks in all arch subdirectories so that
`make distclean` completes successfully.