WIP: Remove choice of inline configuration
Now all functions are `inline`, instead of the common configuration of
`static inline` (named eager inlining).  The functions are given
external linkage in -llxrt which is now always included in
`rtai-config --lxrt-ldflags`.

There is an important reason for all that has happened in this commit,
and that is the fact that `inline` is a keyword that is standardized in
C99.  Before that, it had been a GNU extensions.  That said, the
semantics of `inline` and its interaction with external linkage are
quite different between the two versions.

Clang already defaults to C99, although not used with RTAI yet and gcc
version 5 will also default to C11.  This commit forces the user to use
a compiler that is compliant with C99 (automatically detected in the
configure script), whose availability is not an issue.  This _would_ be
noticed when rebuilding programs that had been using the default gcc C
version (gnu89).  They should now get an error requesting C99 features
to be turned on.  Using `rtai-config --cc`, this option would be
automatically set.

---

For reference, the motivation for this change has been the following.
In C99, it is illegal to call a "static inline" function from a
non-static inline function.  This means that either inlining needs to be
turned off during configuration of RTAI, or all inline users of it also
made static.  For those functions to provide an external linkage as
well, tricks such as the one in RTAI itself (with RTAI_PROTO macro) are
required.  Without this change, that kind of hack would need to be
reimplemented by every library that is built upon RTAI.

---

This commit has the added benefit that some functions defined by
RTAI_PROTO, that with eager inlining would turn into `static inline`
functions, are now just `inline`.  This means that the compiler can
choose to either inline them, or use the external reference.  Some of
the LXRT functions in RTAI are quite large and in such situations,
the compiler will now likely choose to call the external reference,
rather than bloat the code with inline copies of the function or
multiple static inline definitions in multiple files.

Signed-off-by: Shahbaz Youssefi <ShabbyX@gmail.com>