Don't complain about C99 support if in C++
The previous commit (00889a2) gave an error if the compiler didn't have
C99 support to make it clear for the user that C99 is required.  With
C++, inline semantics are similar to C99, but the check had failed.
This commit makes sure the commit is not issued when building C++
applications.

Signed-off-by: Shahbaz Youssefi <ShabbyX@gmail.com>